Add files via upload
Mini Project (RollingDiceGUI  Game)
Java Swing application that creates an instance of the Rolling Dice GUI class and sets it to be visible. This is a common structure for Swing applications. Provides methods for loading and updating images in a Swing application. Represents a Swing GUI application for rolling two dice. 
Tool Used:  IntelliJ IDEA (Java).